requirement.txt
streamlit
pandas
scikit-learn
numpy
Pillow
tensorflow
transformers
textblob
xlsxwriter
huggingface_hub